Update to 8.7

+++ dlls/localspl/provider.c
@@ -237,6 +237,7 @@ typedef struct {
     WCHAR *port;
     WCHAR *print_proc;
     WCHAR *datatype;
+    DWORD attributes;
 
     CRITICAL_SECTION jobs_cs;
     struct list jobs;
@@ -546,6 +547,17 @@ static WCHAR * reg_query_value(HKEY key, const WCHAR *name)
     return ret;
 }
 
+static DWORD reg_query_dword(HKEY hkey, const WCHAR *name)
+{
+    DWORD type, val, size = sizeof(size);
+
+    if (RegQueryValueExW(hkey, name, 0, &type, (BYTE*)&val, &size))
+        return 0;
+    if (type != REG_DWORD)
+        return 0;
+    return val;
+}
+
 static printer_info_t* get_printer_info(const WCHAR *name)
 {
     HKEY hkey, hprinter = NULL;
@@ -582,6 +594,7 @@ static printer_info_t* get_printer_info(const WCHAR *name)
     info->port = reg_query_value(hprinter, L"Port");
     info->print_proc = reg_query_value(hprinter, L"Print Processor");
     info->datatype = reg_query_value(hprinter, L"Datatype");
+    info->attributes = reg_query_dword(hprinter, L"Attributes");
     RegCloseKey(hprinter);
 
     if (!info->name || !info->port || !info->print_proc || !info->datatype)
@@ -3277,10 +3290,124 @@ static BOOL WINAPI fpAddJob(HANDLE hprinter, DWORD level, BYTE *data, DWORD size
     return TRUE;
 }
 
+typedef struct {
+    HMODULE hmod;
+    WCHAR *name;
+    BOOL (WINAPI *enum_datatypes)(WCHAR *, WCHAR *, DWORD,
+            BYTE *, DWORD, DWORD *, DWORD *);
+    HANDLE (WINAPI *open)(WCHAR *, PRINTPROCESSOROPENDATA *);
+    BOOL (WINAPI *print)(HANDLE, WCHAR *);
+    BOOL (WINAPI *close)(HANDLE);
+} printproc_t;
+
+static printproc_t * print_proc_load(const WCHAR *name)
+{
+    WCHAR *reg_path, path[2 * MAX_PATH];
+    printproc_t *ret;
+    DWORD size, len;
+    LSTATUS status;
+    HKEY hkey;
+
+    size = sizeof(fmt_printprocessorsW) +
+        (wcslen(env_arch.envname) + wcslen(name)) * sizeof(WCHAR);
+    reg_path = malloc(size);
+    if (!reg_path)
+        return NULL;
+    swprintf(reg_path, size / sizeof(WCHAR), fmt_printprocessorsW, env_arch.envname);
+    wcscat(reg_path, name);
+
+    status = RegOpenKeyW(HKEY_LOCAL_MACHINE, reg_path, &hkey);
+    free(reg_path);
+    if (status != ERROR_SUCCESS)
+        return NULL;
+
+    if (!fpGetPrintProcessorDirectory(NULL, NULL, 1, (BYTE *)path, sizeof(path), &size))
+    {
+        RegCloseKey(hkey);
+        return NULL;
+    }
+    len = size / sizeof(WCHAR);
+    path[len - 1] = '\\';
+
+    size = sizeof(path) - len * sizeof(WCHAR);
+    status = RegQueryValueExW(hkey, L"Driver", NULL, NULL, (BYTE *)(path + len), &size);
+    RegCloseKey(hkey);
+    if (status != ERROR_SUCCESS)
+        return NULL;
+
+    ret = malloc(sizeof(*ret));
+    if (!ret)
+        return NULL;
+
+    TRACE("loading print processor: %s\n", debugstr_w(path));
+
+    ret->hmod = LoadLibraryW(path);
+    if (!ret->hmod)
+    {
+        free(ret);
+        return NULL;
+    }
+
+    ret->enum_datatypes = (void *)GetProcAddress(ret->hmod, "EnumPrintProcessorDatatypesW");
+    ret->open = (void *)GetProcAddress(ret->hmod, "OpenPrintProcessor");
+    ret->print = (void *)GetProcAddress(ret->hmod, "PrintDocumentOnPrintProcessor");
+    ret->close = (void *)GetProcAddress(ret->hmod, "ClosePrintProcessor");
+    if (!ret->enum_datatypes || !ret->open || !ret->print || !ret->close)
+    {
+        FreeLibrary(ret->hmod);
+        free(ret);
+        return NULL;
+    }
+
+    ret->name = wcsdup(name);
+    return ret;
+}
+
+static BOOL print_proc_check_datatype(printproc_t *pp, const WCHAR *datatype)
+{
+    DATATYPES_INFO_1W *types;
+    DWORD size, no, i;
+
+    if (!datatype)
+        return FALSE;
+
+    pp->enum_datatypes(NULL, pp->name, 1, NULL, 0, &size, &no);
+    if (GetLastError() != ERROR_INSUFFICIENT_BUFFER)
+        return FALSE;
+
+    types = malloc(size);
+    if (!types)
+        return FALSE;
+
+    if (!pp->enum_datatypes(NULL, pp->name, 1, (BYTE *)types, size, &size, &no))
+    {
+        free(types);
+        return FALSE;
+    }
+
+    for (i = 0; i < no; i++)
+    {
+        if (!wcscmp(types[i].pName, datatype))
+            break;
+    }
+    free(types);
+    return i < no;
+}
+
+static void print_proc_unload(printproc_t *pp)
+{
+    FreeLibrary(pp->hmod);
+    free(pp->name);
+    free(pp);
+}
+
 static DWORD WINAPI fpStartDocPrinter(HANDLE hprinter, DWORD level, BYTE *doc_info)
 {
     printer_t *printer = (printer_t *)hprinter;
     DOC_INFO_1W *info = (DOC_INFO_1W *)doc_info;
+    BOOL datatype_valid;
+    WCHAR *datatype;
+    printproc_t *pp;
 
     TRACE("(%p %ld %p {pDocName = %s, pOutputFile = %s, pDatatype = %s})\n",
             hprinter, level, doc_info, debugstr_w(info->pDocName),
@@ -3315,6 +3442,40 @@ static DWORD WINAPI fpStartDocPrinter(HANDLE hprinter, DWORD level, BYTE *doc_in
     if (printer->doc)
     {
         SetLastError(ERROR_INVALID_PRINTER_STATE);
+        return 0;
+    }
+
+    if (info->pDatatype)
+        datatype = info->pDatatype;
+    else if (printer->datatype)
+        datatype = printer->datatype;
+    else
+        datatype = printer->info->datatype;
+
+    if (!datatype || ((printer->info->attributes & PRINTER_ATTRIBUTE_RAW_ONLY) &&
+                wcsicmp(datatype, L"RAW")))
+    {
+        TRACE("non RAW datatype specified on RAW-only printer (%s)\n", debugstr_w(datatype));
+        SetLastError(ERROR_INVALID_DATATYPE);
+        return 0;
+    }
+
+    pp = print_proc_load(printer->info->print_proc);
+    if (!pp)
+    {
+        WARN("failed to load %s print processor\n", debugstr_w(printer->info->print_proc));
+        pp = print_proc_load(L"winprint");
+    }
+    if (!pp)
+        return 0;
+
+    datatype_valid = print_proc_check_datatype(pp, datatype);
+    print_proc_unload(pp);
+    if (!datatype_valid)
+    {
+        TRACE("%s datatype not supported by %s\n", debugstr_w(datatype),
+                debugstr_w(printer->info->print_proc));
+        SetLastError(ERROR_INVALID_DATATYPE);
         return 0;
     }
 
@@ -3557,117 +3718,6 @@ static BOOL WINAPI fpGetJob(HANDLE hprinter, DWORD job_id, DWORD level,
     return ret;
 }
 
-typedef struct {
-    HMODULE hmod;
-    WCHAR *name;
-    BOOL (WINAPI *enum_datatypes)(WCHAR *, WCHAR *, DWORD,
-            BYTE *, DWORD, DWORD *, DWORD *);
-    HANDLE (WINAPI *open)(WCHAR *, PRINTPROCESSOROPENDATA *);
-    BOOL (WINAPI *print)(HANDLE, WCHAR *);
-    BOOL (WINAPI *close)(HANDLE);
-} printproc_t;
-
-static printproc_t * print_proc_load(const WCHAR *name)
-{
-    WCHAR *reg_path, path[2 * MAX_PATH];
-    printproc_t *ret;
-    DWORD size, len;
-    LSTATUS status;
-    HKEY hkey;
-
-    size = sizeof(fmt_printprocessorsW) +
-        (wcslen(env_arch.envname) + wcslen(name)) * sizeof(WCHAR);
-    reg_path = malloc(size);
-    if (!reg_path)
-        return NULL;
-    swprintf(reg_path, size / sizeof(WCHAR), fmt_printprocessorsW, env_arch.envname);
-    wcscat(reg_path, name);
-
-    status = RegOpenKeyW(HKEY_LOCAL_MACHINE, reg_path, &hkey);
-    free(reg_path);
-    if (status != ERROR_SUCCESS)
-        return NULL;
-
-    if (!fpGetPrintProcessorDirectory(NULL, NULL, 1, (BYTE *)path, sizeof(path), &size))
-    {
-        RegCloseKey(hkey);
-        return NULL;
-    }
-    len = size / sizeof(WCHAR);
-    path[len - 1] = '\\';
-
-    size = sizeof(path) - len * sizeof(WCHAR);
-    status = RegQueryValueExW(hkey, L"Driver", NULL, NULL, (BYTE *)(path + len), &size);
-    RegCloseKey(hkey);
-    if (status != ERROR_SUCCESS)
-        return NULL;
-
-    ret = malloc(sizeof(*ret));
-    if (!ret)
-        return NULL;
-
-    TRACE("loading print processor: %s\n", debugstr_w(path));
-
-    ret->hmod = LoadLibraryW(path);
-    if (!ret->hmod)
-    {
-        free(ret);
-        return NULL;
-    }
-
-    ret->enum_datatypes = (void *)GetProcAddress(ret->hmod, "EnumPrintProcessorDatatypesW");
-    ret->open = (void *)GetProcAddress(ret->hmod, "OpenPrintProcessor");
-    ret->print = (void *)GetProcAddress(ret->hmod, "PrintDocumentOnPrintProcessor");
-    ret->close = (void *)GetProcAddress(ret->hmod, "ClosePrintProcessor");
-    if (!ret->enum_datatypes || !ret->open || !ret->print || !ret->close)
-    {
-        FreeLibrary(ret->hmod);
-        free(ret);
-        return NULL;
-    }
-
-    ret->name = wcsdup(name);
-    return ret;
-}
-
-static BOOL print_proc_check_datatype(printproc_t *pp, const WCHAR *datatype)
-{
-    DATATYPES_INFO_1W *types;
-    DWORD size, no, i;
-
-    if (!datatype)
-        return FALSE;
-
-    pp->enum_datatypes(NULL, pp->name, 1, NULL, 0, &size, &no);
-    if (GetLastError() != ERROR_INSUFFICIENT_BUFFER)
-        return FALSE;
-
-    types = malloc(size);
-    if (!types)
-        return FALSE;
-
-    if (!pp->enum_datatypes(NULL, pp->name, 1, (BYTE *)types, size, &size, &no))
-    {
-        free(types);
-        return FALSE;
-    }
-
-    for (i = 0; i < no; i++)
-    {
-        if (!wcscmp(types[i].pName, datatype))
-            break;
-    }
-    free(types);
-    return i < no;
-}
-
-static void print_proc_unload(printproc_t *pp)
-{
-    FreeLibrary(pp->hmod);
-    free(pp->name);
-    free(pp);
-}
-
 static BOOL WINAPI fpScheduleJob(HANDLE hprinter, DWORD job_id)
 {
     printer_t *printer = (printer_t *)hprinter;
@@ -3766,7 +3816,8 @@ static BOOL WINAPI fpScheduleJob(HANDLE hprinter, DWORD job_id)
     pp->close(hpp);
     print_proc_unload(pp);
 
-    DeleteFileW(job->filename);
+    if (!(printer->info->attributes & PRINTER_ATTRIBUTE_KEEPPRINTEDJOBS))
+        DeleteFileW(job->filename);
     free_job(job);
     LeaveCriticalSection(&printer->info->jobs_cs);
     return ret;
